+++ docs/changes.rst
@@ -1785,7 +1785,7 @@ Released on November 16th 2012.
 * Compatibility with PostgreSQL database backend.
 * Removes languages removed in upstream git repository.
 * Improved quality checks processing.
-* Added new checks (BB code, XML markup and newlines).
+* Added new checks (BBCode, XML markup and newlines).
 * Support for optional rebasing instead of merge.
 * Possibility to relocate Weblate (for example to run it under /weblate path).
 * Support for manually choosing file type in case autodetection fails.